+++ src/main/java/entity/Room.java
@@ -1,0 +1,4 @@
+package entity;
+
+public class Room {
+}
