--- src/main/java/org/hospitality/entity/Room.java
@@ -1,4 +1,0 @@
-package org.hospitality.entity;
-
-public class Room {
-}
